Split ParserUtil class
