+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -1,39 +1,11 @@
 package seedu.address.logic.parser;
 
-import static java.util.Objects.requireNonNull;
-
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
 import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.MissingPrefixesException;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.consultation.ConsultationDescription;
-import seedu.address.model.consultation.ConsultationModule;
-import seedu.address.model.consultation.ConsultationName;
-import seedu.address.model.consultation.ConsultationTimeslot;
-import seedu.address.model.consultation.ConsultationVenue;
-import seedu.address.model.reminder.ReminderDeadline;
-import seedu.address.model.reminder.ReminderDescription;
-import seedu.address.model.reminder.ReminderName;
-import seedu.address.model.student.Address;
-import seedu.address.model.student.Attendance;
-import seedu.address.model.student.Email;
-import seedu.address.model.student.Grade;
-import seedu.address.model.student.ID;
-import seedu.address.model.student.Name;
-import seedu.address.model.student.Participation;
-import seedu.address.model.student.Phone;
-import seedu.address.model.student.Telegram;
-import seedu.address.model.tag.Tag;
-import seedu.address.model.tutorial.TutorialDay;
-import seedu.address.model.tutorial.TutorialModule;
-import seedu.address.model.tutorial.TutorialName;
-import seedu.address.model.tutorial.TutorialTimeslot;
-import seedu.address.model.tutorial.TutorialVenue;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
@@ -71,351 +43,4 @@ public class ParserUtil {
         }
     }
 
-    /**
-     * Parses a {@code String name} into a {@code Name}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code name} is invalid.
-     */
-    public static Name parseName(String name) throws ParseException {
-        requireNonNull(name);
-        String trimmedName = name.trim();
-        if (!Name.isValidName(trimmedName)) {
-            throw new ParseException(Name.MESSAGE_CONSTRAINTS);
-        }
-        return new Name(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String phone} into a {@code Phone}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code phone} is invalid.
-     */
-    public static Phone parsePhone(String phone) throws ParseException {
-        requireNonNull(phone);
-        String trimmedPhone = phone.trim();
-        if (!Phone.isValidPhone(trimmedPhone)) {
-            throw new ParseException(Phone.MESSAGE_CONSTRAINTS);
-        }
-        return new Phone(trimmedPhone);
-    }
-
-    /**
-     * Parses a {@code String address} into an {@code Address}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code address} is invalid.
-     */
-    public static Address parseAddress(String address) throws ParseException {
-        requireNonNull(address);
-        String trimmedAddress = address.trim();
-        if (!Address.isValidAddress(trimmedAddress)) {
-            throw new ParseException(Address.MESSAGE_CONSTRAINTS);
-        }
-        return new Address(trimmedAddress);
-    }
-
-    /**
-     * Parses a {@code String email} into an {@code Email}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code email} is invalid.
-     */
-    public static Email parseEmail(String email) throws ParseException {
-        requireNonNull(email);
-        String trimmedEmail = email.trim();
-        if (!Email.isValidEmail(trimmedEmail)) {
-            throw new ParseException(Email.MESSAGE_CONSTRAINTS);
-        }
-        return new Email(trimmedEmail);
-    }
-
-    /**
-     * Parses a {@code String tag} into a {@code Tag}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code tag} is invalid.
-     */
-    public static Tag parseTag(String tag) throws ParseException {
-        requireNonNull(tag);
-        String trimmedTag = tag.trim();
-        if (!Tag.isValidTagName(trimmedTag)) {
-            throw new ParseException(Tag.MESSAGE_CONSTRAINTS);
-        }
-        return new Tag(trimmedTag);
-    }
-
-    /**
-     * Parses {@code Collection<String> tags} into a {@code Set<Tag>}.
-     */
-    public static Set<Tag> parseTags(Collection<String> tags) throws ParseException {
-        requireNonNull(tags);
-        final Set<Tag> tagSet = new HashSet<>();
-        for (String tagName : tags) {
-            tagSet.add(parseTag(tagName));
-        }
-        return tagSet;
-    }
-
-    /**
-     * Parses a {@code String name} into a {@code ReminderName}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code ReminderName} is invalid.
-     */
-    public static ReminderName parseReminderName(String name) throws ParseException {
-        requireNonNull(name);
-        String trimmedName = name.trim();
-        if (!ReminderName.isValidName(trimmedName)) {
-            throw new ParseException(ReminderName.MESSAGE_CONSTRAINTS);
-        }
-        return new ReminderName(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String deadline} into a {@code ReminderDeadline}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code ReminderDeadline} is invalid.
-     */
-    public static ReminderDeadline parseReminderDeadline(String deadline) throws ParseException {
-        requireNonNull(deadline);
-        String trimmedDeadline = deadline.trim();
-        if (!ReminderDeadline.isValidTimeslot(trimmedDeadline)) {
-            throw new ParseException(ReminderDeadline.MESSAGE_CONSTRAINTS);
-        }
-        return new ReminderDeadline(trimmedDeadline);
-    }
-
-    /**
-     * Parses a {@code String description} into a {@code ReminderDescription}.
-     * Leading and trailing whitespaces will be trimmed.
-     */
-    public static ReminderDescription parseReminderDescription(String description) {
-        String trimmedDescription = description.trim();
-        return new ReminderDescription(trimmedDescription);
-    }
-
-    /**
-     * Parses a {@code String name} into a {@code TutorialName}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code name} is invalid.
-     */
-    public static TutorialName parseTutorialName(String name) throws ParseException {
-        requireNonNull(name);
-        String trimmedName = name.trim();
-        if (!TutorialName.isValidName(trimmedName)) {
-            throw new ParseException(TutorialName.MESSAGE_CONSTRAINTS);
-        }
-        return new TutorialName(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String moduleName} into a {@code TutorialModule}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code moduleName} is invalid.
-     */
-    public static TutorialModule parseTutorialModule(String moduleName) throws ParseException {
-        requireNonNull(moduleName);
-        String trimmedName = moduleName.trim();
-        if (!TutorialModule.isValidModule(trimmedName)) {
-            throw new ParseException(TutorialModule.MESSAGE_CONSTRAINTS);
-        }
-        return new TutorialModule(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String venue} into a {@code TutorialVenue}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code venue} is invalid.
-     */
-    public static TutorialVenue parseTutorialVenue(String venue) throws ParseException {
-        requireNonNull(venue);
-        String trimmedName = venue.trim();
-        if (!TutorialVenue.isValidVenue(trimmedName)) {
-            throw new ParseException(TutorialVenue.MESSAGE_CONSTRAINTS);
-        }
-        return new TutorialVenue(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String timeslot} into a {@code TutorialTimeslot}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code timeslot} is invalid.
-     */
-    public static TutorialTimeslot parseTutorialTimeslot(String timeslot) throws ParseException {
-        requireNonNull(timeslot);
-        String trimmedName = timeslot.trim();
-        if (!TutorialTimeslot.isValidTimeslot(trimmedName)) {
-            throw new ParseException(TutorialTimeslot.MESSAGE_CONSTRAINTS);
-        }
-        if (!TutorialTimeslot.isValidDuration(trimmedName)) {
-            throw new ParseException(TutorialTimeslot.MESSAGE_INVALID_DURATION);
-        }
-        return new TutorialTimeslot(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String day} into a {@code TutorialDay}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code day} is invalid.
-     */
-    public static TutorialDay parseTutorialDay(String day) throws ParseException {
-        requireNonNull(day);
-        String trimmedValue = day.trim();
-        if (!TutorialDay.isValidDay(trimmedValue)) {
-            throw new ParseException(TutorialDay.MESSAGE_CONSTRAINTS);
-        }
-        return new TutorialDay(trimmedValue);
-    }
-
-    /**
-     * Parses a {@code String name} into a {@code TutorialName}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code name} is invalid.
-     */
-    public static ConsultationName parseConsultationName(String name) throws ParseException {
-        requireNonNull(name);
-        String trimmedName = name.trim();
-        if (!ConsultationName.isValidName(trimmedName)) {
-            throw new ParseException(ConsultationName.MESSAGE_CONSTRAINTS);
-        }
-        return new ConsultationName(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String moduleName} into a {@code ConsultationModule}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code moduleName} is invalid.
-     */
-    public static ConsultationModule parseConsultationModule(String moduleName) throws ParseException {
-        requireNonNull(moduleName);
-        String trimmedName = moduleName.trim();
-        if (!ConsultationModule.isValidModule(trimmedName)) {
-            throw new ParseException(ConsultationModule.MESSAGE_CONSTRAINTS);
-        }
-        return new ConsultationModule(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String venue} into a {@code ConsultationVenue}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code venue} is invalid.
-     */
-    public static ConsultationVenue parseConsultationVenue(String venue) throws ParseException {
-        requireNonNull(venue);
-        String trimmedName = venue.trim();
-        if (!ConsultationVenue.isValidVenue(trimmedName)) {
-            throw new ParseException(ConsultationVenue.MESSAGE_CONSTRAINTS);
-        }
-        return new ConsultationVenue(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String timeslot} into a {@code ConsultationTimeslot}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code timeslot} is invalid.
-     */
-    public static ConsultationTimeslot parseConsultationTimeslot(String timeslot) throws ParseException {
-        requireNonNull(timeslot);
-        String trimmedName = timeslot.trim();
-        if (!ConsultationTimeslot.isValidTimeslot(trimmedName)) {
-            throw new ParseException(ConsultationTimeslot.MESSAGE_CONSTRAINTS);
-        }
-        if (!ConsultationTimeslot.isValidDuration(trimmedName)) {
-            throw new ParseException(ConsultationTimeslot.MESSAGE_INVALID_DURATION);
-        }
-        return new ConsultationTimeslot(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String name} into a {@code TutorialName}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code name} is invalid.
-     */
-    public static ConsultationDescription parseConsultationDescription(String name) throws ParseException {
-        requireNonNull(name);
-        String trimmedName = name.trim();
-        if (!ConsultationDescription.isValidDescription(trimmedName)) {
-            throw new ParseException(ConsultationName.MESSAGE_CONSTRAINTS);
-        }
-        return new ConsultationDescription(trimmedName);
-    }
-
-    /**
-     * Parses a {@code String id} into a {@code ID}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code id} is invalid.
-     */
-    public static ID parseId(String id) throws ParseException {
-        requireNonNull(id);
-        String trimmedId = id.trim();
-        if (!ID.isValidId(trimmedId)) {
-            throw new ParseException(ID.MESSAGE_CONSTRAINTS);
-        }
-        return new ID(trimmedId);
-    }
-
-    /**
-     * Parses a {@code String telegram} into a {@code Telegram}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code telegram} is invalid.
-     */
-    public static Telegram parseTelegram(String telegram) throws ParseException {
-        requireNonNull(telegram);
-        String trimmedTelegram = telegram.trim();
-        if (!Telegram.isValidTelegram(trimmedTelegram)) {
-            throw new ParseException(Telegram.MESSAGE_CONSTRAINTS);
-        }
-        return new Telegram(trimmedTelegram);
-    }
-
-    /**
-     * Parses a {@code String grade} into a {@code Grade}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code grade} is invalid.
-     */
-    public static Grade parseGrade(String grade) {
-        requireNonNull(grade);
-        String trimmedGrade = grade.trim();
-        return new Grade(trimmedGrade);
-    }
-
-    /**
-     * Parses a {@code String attendance} into a {@code Attendance}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code attendance} is invalid.
-     */
-    public static Attendance parseAttendance(String attendance) {
-        requireNonNull(attendance);
-        String trimmedAttendance = attendance.trim();
-        return new Attendance(trimmedAttendance);
-    }
-
-    /**
-     * Parses a {@code String participation} into a {@code Participation}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code participation} is invalid.
-     */
-    public static Participation parseParticipation(String participation) {
-        requireNonNull(participation);
-        String trimmedParticipation = participation.trim();
-        return new Participation(trimmedParticipation);
-    }
 }
